Removed Scala 2.11 support and extracted version specific classes to own source folder
